Added component's name to sidebar
