+++ client/src/sass/all.scss
@@ -1,73 +1,15 @@
 @import '../components/View/view';
 @import '../components/Nav/nav';
-@import '../components/SideBar/sidebar';
 
 
-html,body{height:100%}
-//.wrap{
-//  width:100%;
-//  height:100%;
-//  position:relative;
-//}
-//
-//.nav{
-//  height:7.5%;
-//  position:fixed;
-//  top:0;
-//  left:0;
-//  width:100%;
-//}
-//
-//.footer{
-//  height:10%;
-//  position:fixed;
-//  bottom:0;
-//  width:100%;
-//  margin-left:14%;
-//  padding: 2.5%;
-//  z-index: -2;
-//}
-//
-//.bodywrap{
-//  margin-top:100px;
-//}
-//
-//.left{
-//  width:15%;
-//  top: 6%;
-//  bottom: 0px;
-//  position: fixed;
-//  z-index: -1;
-//}
-//
-//.left, .center, .bodywrap{
-//  height:100%
-//}
-//
-//
-//.left{
-//  left:0
-//}
-//
-//.left{
-//  background-color: #038ba7
-//}
-//.footer{background-color: #414141}
-//
-//.floatcenter {
-//  margin: auto;
-//  width: 50%;
-//  display: inline-block;
-//}
-//.floatright {
-//  float:right;
-//  padding-right: 5%;
-//}
+html,body{
+  height: 100%
+}
 
 .row {
   margin-bottom: 0;
 }
 
 .center-content{
-  margin:8% 15%;
+  margin: 8% 15%;
 }
