+++ package.json
@@ -40,6 +40,7 @@
     "grunt-cli": "^1.2.0",
     "grunt-pg": "https://github.com/watsoncj/grunt-pg.git",
     "knex": "^0.13.0",
+    "lodash": "^4.17.4",
     "material-ui": "^0.18.1",
     "morgan": "^1.8.1",
     "node-sass": "^4.5.3",
